so, get everything back to a working state:
1. make the code re: generic types no longer stateful. change the name of generic types before generation so that the "most-recently-generated-type" can't be looked up from the basename (relax, no mangledName-based-lookup)
2. propagate the necessity of the generic type map to more places in the code; basically, if you know about generic types, then you are required to pass on that knowledge to stuff that you touch (if they need it)
3. enabled warnings on MSVC, and fix a couple of those signedness things.

+++ source/frontend/errors.cpp
@@ -232,7 +232,7 @@ std::string __error_gen_backtrace(const HighlightOptions& ops)
 
 			seen.push_back(e->loc);
 
-			int skip = numlocs - done - i;
+			int64_t skip = (int64_t) numlocs - done - (int64_t) i;
 			if(done == 2 && i - MAX_BACKTRACE_DEPTH > 1 && skip > 0)
 			{
 				ret += strprintf("... skipping %d intermediarie%s...\n\n", skip, skip == 1 ? "" : "s");
